Random mouse movement
Generate random coords from the map, and use the go_to_goal to move the robot to the position.
Made it dynamic using the metadata from the map.
Uses stdr from the cat_and_mouse_world package. This stdr version implements the names of the nodes so it is not compatible with stdr