Closer initial robots
Moved the initial cat & mouse to only 0.5.
Tried to improve cat movement.
go_to_goal has a faster movement as they get closer.
Moved the set rate of mouse movement to init & made it sleep at the begining to try improve the initial time of the cat.

The initial movement of the cat as a weird behaviour. It has the point it should move to but it stays still.